+++ src/com/gazman/quadratic_sieve/data/VectorWorkData.java
@@ -4,10 +4,10 @@ import java.math.BigInteger;
 
 public class VectorWorkData {
     public final PolynomialData polynomialData;
-    public final BigInteger localX;
+    public final long localX;
     public final BigInteger c;
 
-    public VectorWorkData(PolynomialData polynomialData, BigInteger localX, BigInteger c) {
+    public VectorWorkData(PolynomialData polynomialData, long localX, BigInteger c) {
         this.polynomialData = polynomialData;
         this.localX = localX;
         this.c = c;
